fix migration file

+++ database/migrations/2024_03_15_162623_create_house_utilities__table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->unsignedBigInteger('utility_id');
             $table->string('image')->nullable();
             $table->integer('price')->nullable();
-            $table->integer('quantity')->nullable();
+            $table->tinyInteger('quantity')->nullable();
             $table->timestamps();
 
             $table->foreign('house_id')
